# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Exakta Laboratories turn taking application",
        "version": "1.0.0",
        "title": "API Turnero",
        "contact": {
            "email": "exakta@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:4000",
    "basePath": "/api",
...
    "paths" : {
...
        "/config": {
            "get": {
                "tags": ["Configuration"],
                "summary" : "Get configuration system",
                "description": "Get configuration system",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/ConfigRes"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": [
                                    {
                                        "_id": "61ead42791ab1e940eb3cf78",
                                        "timer": 2000,
                                        "createdAt": "2022-01-21T15:41:27.325Z",
                                        "updatedAt": "2022-01-21T18:38:43.536Z",
                                        "__v": 0
                                    }
                                ],
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            },
            "post": {
                "tags": ["Configuration"],
                "summary" : "Create a data of configuration system",
                "description": "Create a data of configuration system",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "configuration",
                        "description": "Configuration object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ConfigReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/ConfigRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 201,
                                "body": {
                                    "_id": "61ead42791ab1e940eb3cf78",
                                    "timer": 2000,
                                    "createdAt": "2022-01-21T15:41:27.325Z",
                                    "updatedAt": "2022-01-21T18:38:43.536Z",
                                    "__v": 0
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            },
            "delete": {
                "tags": ["Configuration"],
                "summary" : "Delete a configuration system",
                "description": "Delete a configuration system",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "object",
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": {
                                    "message": "Successfully delete."
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 404,
                                "body": {},
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            }
        }
    }, 
    "definitions": {
...
        "InvalidRes": {
            "type": "object",
            "properties": {
                "message":  {
                    "type": "string"
                }
            }
        },
...
        "ConfigRes": {
            "type": "object",
            "properties": {
                "_id":  {
                    "type": "string"
                },
                "timer":  {
                    "type": "number"
                },
                "createdAt":  {
                    "type": "string"
                },
                "updatedAt":  {
                    "type": "string"
                },
                "__v":  {
                    "type": "number"
                }
            }
        },
        "ConfigReq": {
            "type": "object",
            "properties": {
                "timer":  {
                    "type": "number"
                }
            },
            "required": [ 
                "timer"
            ],
            "example": {
                "timer": 5000
            } 
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "auth",
            "in": "header"
        }
    }
}